init conversion to robot
