+++ urdf2webots/importer.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-"""URDF files to Webots PROTO converter."""
+"""URDF files to Webots robot converter."""
 
 
 import os
@@ -9,7 +9,7 @@ import optparse
 import re
 import sys
 import urdf2webots.parserURDF
-import urdf2webots.writeProto
+import urdf2webots.writeRobot
 from xml.dom import minidom
 
 try:
@@ -46,12 +46,27 @@ def mkdirSafe(directory):
             print('Directory "' + directory + '" already exists!')
 
 
-def convert2urdf(inFile, outFile=None, normal=False, boxCollision=False,
+def convert2urdf(inFile, worldFile, robotName='', normal=False, boxCollision=False,
                  disableMeshOptimization=False, enableMultiFile=False,
-                 staticBase=False, toolSlot=None, initRotation='0 1 0 0',
-                 initPos=None, linkToDef=False, jointToDef=False):
-    if not os.path.exists(inFile):
+                 staticBase=False, initTranslation='0 0 0',
+                 initRotation='0 1 0 0', initPos=None, linkToDef=False,
+                 jointToDef=False):
+    if not inFile:
+        sys.exit('Input file not specified (should be specified with the "--input" argument).')
+    if not os.path.isfile(inFile):
         sys.exit('Input file "%s" does not exists.' % inFile)
+    if not inFile.endswith('.urdf'):
+        sys.exit('"%s" is not an urdf file.' % inFile)
+
+    if not worldFile:
+        sys.exit('World file not specified (should be specified with the "--worldFile" argument).')
+    if not os.path.isfile(worldFile):
+        sys.exit('World file "%s" does not exists.' % inFile)
+    if not worldFile.endswith('.wbt'):
+        sys.exit('"%s" is not an world file.' % worldFile)
+
+    if not type(initTranslation) == str or len(initTranslation.split()) != 3:
+        sys.exit('--translation argument is not valid. Has to be of Type = str and contain 3 values.')
     if not type(initRotation) == str or len(initRotation.split()) != 4:
         sys.exit('--rotation argument is not valid. Has to be of Type = str and contain 4 values.')
     if initPos is not None:
@@ -63,12 +78,10 @@ def convert2urdf(inFile, outFile=None, normal=False, boxCollision=False,
                      'Example: --init-pos="1.0, 2, -0.4"')
 
     urdf2webots.parserURDF.disableMeshOptimization = disableMeshOptimization
-    urdf2webots.writeProto.enableMultiFile = enableMultiFile
-    urdf2webots.writeProto.staticBase = staticBase
-    urdf2webots.writeProto.toolSlot = toolSlot
-    urdf2webots.writeProto.initPos = initPos
-    urdf2webots.writeProto.linkToDef = linkToDef
-    urdf2webots.writeProto.jointToDef = jointToDef
+    urdf2webots.writeRobot.enableMultiFile = enableMultiFile
+    urdf2webots.writeRobot.initPos = initPos
+    urdf2webots.writeRobot.linkToDef = linkToDef
+    urdf2webots.writeRobot.jointToDef = jointToDef
 
     with open(inFile, 'r') as file:
         inPath = os.path.dirname(os.path.abspath(inFile))
@@ -98,33 +111,29 @@ def convert2urdf(inFile, outFile=None, normal=False, boxCollision=False,
 
         for child in domFile.childNodes:
             if child.localName == 'robot':
-                if outFile:
-                    if os.path.splitext(os.path.basename(outFile))[1] == '.proto':
-                        robotName = os.path.splitext(os.path.basename(outFile))[0]
-                        outputFile = outFile
-                    else:
-                        # treat outFile as directory and construct filename
-                        robotName = convertLUtoUN(urdf2webots.parserURDF.getRobotName(child))  # capitalize
-                        outputFile = os.path.join(outFile, robotName + '.proto')
-                else:
+                if robotName == '':
                     robotName = convertLUtoUN(urdf2webots.parserURDF.getRobotName(child))  # capitalize
-                    outputFile = outFile if outFile else robotName + '.proto'
-                urdf2webots.writeProto.robotNameMain = robotName
+                urdf2webots.writeRobot.robotName = robotName
 
                 urdf2webots.parserURDF.robotName = robotName  # pass robotName
-                mkdirSafe(outputFile.replace('.proto', '') + '_textures')  # make a dir called 'x_textures'
+                mkdirSafe(worldFile.replace('.wbt', '') + '_textures')  # make a dir called 'x_textures'
 
                 if enableMultiFile:
-                    mkdirSafe(outputFile.replace('.proto', '') + '_meshes')  # make a dir called 'x_meshes'
-                    urdf2webots.writeProto.meshFilesPath = outputFile.replace('.proto', '') + '_meshes'
+                    mkdirSafe(worldFile.replace('.wbt', '') + '_meshes')  # make a dir called 'x_meshes'
+                    urdf2webots.writeRobot.meshFilesPath = worldFile.replace('.wbt', '') + '_meshes'
 
                 robot = child
-                protoFile = open(outputFile, 'w')
-                urdf2webots.writeProto.header(protoFile, inFile, robotName)
+                worldFile = open(worldFile, 'a')
+                urdf2webots.writeRobot.header(worldFile, inFile)
                 linkElementList = []
                 jointElementList = []
                 for child in robot.childNodes:
-                    if child.localName == 'link':
+                    # Check if the URDF requires a static base
+                    if child.localName == 'webots':
+                        for webotsChild in child.childNodes:
+                            if webotsChild.localName == 'staticBase':
+                                urdf2webots.writeRobot.staticBase = webotsChild.firstChild.nodeValue
+                    elif child.localName == 'link':
                         linkElementList.append(child)
                     elif child.localName == 'joint':
                         jointElementList.append(child)
@@ -167,7 +176,6 @@ def convert2urdf(inFile, outFile=None, normal=False, boxCollision=False,
                                 rootLink = previousRootLink
                                 break
 
-
                         print('Root link: ' + rootLink.name)
                         break
 
@@ -181,11 +189,10 @@ def convert2urdf(inFile, outFile=None, normal=False, boxCollision=False,
                               urdf2webots.parserURDF.Lidar.list)
                 print('There are %d links, %d joints and %d sensors' % (len(linkList), len(jointList), len(sensorList)))
 
-                urdf2webots.writeProto.declaration(protoFile, robotName, initRotation)
-                urdf2webots.writeProto.URDFLink(protoFile, rootLink, 1, parentList, childList, linkList, jointList,
-                                                sensorList, boxCollision=boxCollision, normal=normal, robot=True)
-                protoFile.write('}\n')
-                protoFile.close()
+                urdf2webots.writeRobot.URDFLink(worldFile, rootLink, 0, parentList,
+                                                childList, linkList, jointList, sensorList, boxCollision=boxCollision,
+                                                normal=normal, robot=True, initTranslation=initTranslation, initRotation=initRotation)
+                worldFile.close()
                 return
     print('Could not read file')
 
@@ -193,8 +200,9 @@ def convert2urdf(inFile, outFile=None, normal=False, boxCollision=False,
 if __name__ == '__main__':
     optParser = optparse.OptionParser(usage='usage: %prog --input=my_robot.urdf [options]')
     optParser.add_option('--input', dest='inFile', default='', help='Specifies the urdf file to convert.')
-    optParser.add_option('--output', dest='outFile', default='', help='Specifies the path and, if ending in ".proto", name '
-                         'of the resulting PROTO file. The filename minus the .proto extension will be the robot name.')
+    optParser.add_option('--worldFile', dest='worldFile', default='', help='Specifies the world file where the result of the '
+                         'conversion will be inserted.')
+    optParser.add_option('--robotName', dest='robotName', default='', help='Specifies the name of the robot.')
     optParser.add_option('--normal', dest='normal', action='store_true', default=False,
                          help='If set, the normals are exported if present in the URDF definition.')
     optParser.add_option('--box-collision', dest='boxCollision', action='store_true', default=False,
@@ -204,12 +212,10 @@ if __name__ == '__main__':
                          'conversion).')
     optParser.add_option('--multi-file', dest='enableMultiFile', action='store_true', default=False,
                          help='If set, the mesh files are exported as separated PROTO files')
-    optParser.add_option('--static-base', dest='staticBase', action='store_true', default=False,
-                         help='If set, the base link will have the option to be static (disable physics)')
-    optParser.add_option('--tool-slot', dest='toolSlot', default=None,
-                         help='Specify the link that you want to add a tool slot too (exact link name from urdf)')
+    optParser.add_option('--translation', dest='initTranslation', default='0 0 0',
+                         help='Set the rotation field of your robot in the world file.')
     optParser.add_option('--rotation', dest='initRotation', default='0 1 0 0',
-                         help='Set the rotation field of your PROTO file.')
+                         help='Set the translation field of your robot in the world file.')
     optParser.add_option('--init-pos', dest='initPos', default=None,
                          help='Set the initial positions of your robot joints. Example: --init-pos="[1.2, 0.5, -1.5]" would '
                          'set the first 3 joints of your robot to the specified values, and leave the rest with their '
@@ -219,5 +225,5 @@ if __name__ == '__main__':
     optParser.add_option('--joint-to-def', dest='jointToDef', action='store_true', default=False,
                          help='If set, urdf joint names are also used as DEF names as well as joint names.')
     options, args = optParser.parse_args()
-    convert2urdf(options.inFile, options.outFile, options.normal, options.boxCollision, options.disableMeshOptimization,
-                 options.enableMultiFile, options.staticBase, options.toolSlot, options.initRotation, options.initPos, options.linkToDef, options.jointToDef)
+    convert2urdf(options.inFile, options.worldFile, options.robotName, options.normal, options.boxCollision, options.disableMeshOptimization,
+                 options.enableMultiFile, options.initTranslation, options.initRotation, options.initPos, options.linkToDef, options.jointToDef)
